+++ localStotageApi.ts
@@ -1,12 +1,14 @@
-const saveResultInLocalStorage = (key, data) => {
+import { Country } from "./types";
+
+const saveResultInLocalStorage = (key: string, data: unknown[]) => {
     localStorage.setItem(key, JSON.stringify(data));
 }
 
-const getResultFromLocalStorage = (key) => {
+const getResultFromLocalStorage = (key: string) => {
     const value = localStorage.getItem(key);
 
     if (value !== null) {
-        return JSON.parse(value);
+        return JSON.parse(value) as Country[];
     }
     return [];
 }
